pyproject.toml
[project]
name = "investing-agent"
version = "0.1.0"
description = "Investing Agent: deterministic, provenance-first valuation and reporting"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Chaorong Chen" }
]
dependencies = [
  "numpy>=1.26",
  "pydantic>=2.6",
  "matplotlib>=3.8",
  "requests>=2.31",
  "openai>=1.0",
  "anthropic>=0.25",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pydantic>=2.7",
  "numpy>=1.26",
  "pandas>=2.2",
  "matplotlib>=3.9",
  "requests>=2.32",
  "python-dateutil>=2.9",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-requests",
  "types-python-dateutil",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 100
respect-gitignore = true

[tool.ruff.lint]
select = [
  "E", "F", "I",  # pycodestyle/pyflakes/isort
  "B",              # flake8-bugbear
  "UP",             # pyupgrade
  "PLC", "PLW",     # pylint (convention/warning subsets)
]
ignore = [
]

[tool.ruff.format]
quote-style = "preserve"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
warn_redundant_casts = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
pretty = true
show_error_codes = true
strict_equality = true
plugins = []
exclude = [
  "^tests/",
]
